Finish the UI implementation

+++ constants.hpp
@@ -1,3 +1,3 @@
-constexpr size_t MEASUREMENT_MEAN_OVER = 10;
-constexpr size_t SCREEN_WIDTH = 80;
-constexpr size_t SCREEN_HEIGHT = 24;
+constexpr int MEASUREMENT_MEAN_OVER = 10;
+constexpr int SCREEN_WIDTH = 80;
+constexpr int SCREEN_HEIGHT = 24;
